requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.3
modal>=0.54.3723
pydantic>=2.5.0
python-dotenv>=1.0.0
google-generativeai>=0.3.1
